Get variable labels from their comment attributes

Usage

codebook(data, cols)

Arguments

data

A tibble.

cols

A tidy variable selections to filter specific columns.

Value

A tibble with the columns: - item_name: The column name. - item_group: First part of the column name, up to an underscore. - item_class: The last class value of an item (e.g. numeric, factor). - item_label: The comment attribute of the column. - value_name: In case a column has numeric attributes, the attribute names - value_label: In case a column has numeric attributes or T/F-attributes, the attribute values. In case a column has a levels attribute, the levels.

Examples

volker::codebook(volker::chatgpt)
#> # A tibble: 94 × 6
#>    item_name     item_group item_class item_label         value_name value_label
#>    <chr>         <chr>      <chr>      <chr>              <chr>      <chr>      
#>  1 case          case       NA         case               NA         NA         
#>  2 sd_age        sd         NA         Age                NA         NA         
#>  3 cg_activities cg         NA         Activities with C… NA         NA         
#>  4 use_private   use        NA         Usage: in private… 1          never      
#>  5 use_private   use        NA         Usage: in private… 2          rarely     
#>  6 use_private   use        NA         Usage: in private… 3          several ti…
#>  7 use_private   use        NA         Usage: in private… 4          several ti…
#>  8 use_private   use        NA         Usage: in private… 5          almost dai…
#>  9 use_work      use        NA         Usage: in profess… 1          never      
#> 10 use_work      use        NA         Usage: in profess… 2          rarely     
#> # ℹ 84 more rows