Output effect sizes for count data

The type of effect size depends on the number of selected columns:

One column: see effect_counts_one (not yet implemented)

Multiple columns: see effect_counts_items (not yet implemented)

One column and one grouping column: see effect_counts_one_grouped

Multiple columns and one grouping column: see effect_counts_items_grouped (not yet implemented)

By default, if you provide two column selections, the second column is treated as categorical. Setting the metric-parameter to TRUE will call the appropriate functions for correlation analysis:

One column and one metric column: see effect_counts_one_cor (not yet implemented)

Multiple columns and one metric column: see effect_counts_items_cor (not yet implemented)

Usage

effect_counts(data, cols, cross = NULL, metric = FALSE, clean = TRUE, ...)

Arguments

data

A data frame.

cols

A tidy column selection, e.g. a single column (without quotes) or multiple columns selected by methods such as starts_with().

cross

Optional, a grouping column. The column name without quotes.

metric

When crossing variables, the cross column parameter can contain categorical or metric values. By default, the cross column selection is treated as categorical data. Set metric to TRUE, to treat it as metric and calculate correlations.

clean

Prepare data by data_clean.

...

Other parameters passed to the appropriate effect function.

Value

A volker tibble.

Examples

library(volker)
data <- volker::chatgpt

effect_counts(data, sd_gender, adopter)
#> 
#> 
#> |Statistic          | Value|
#> |:------------------|-----:|
#> |Number of cases    |   101|
#> |Cramer's V         |  0.09|
#> |Degrees of freedom |      |
#> |Chi-squared        | 13.48|
#> |p value            | 0.022|
#> |stars              |     *|