A reduced skimmer for metric variables Returns a five point summary, mean and sd, items count and alpha for scales added by idx_add()

A reduced skimmer for metric variables Returns a five point summary, mean and sd, items count and alpha for scales added by idx_add()

Usage

skim_metrics(data, ..., .data_name = NULL)

Value

A skimmer, see skim_with

Examples

library(volker)
data <- volker::chatgpt

skim_metrics(data)
#> ── Data Summary ────────────────────────
#>                            Values
#> Name                       data  
#> Number of rows             101   
#> Number of columns          19    
#> _______________________          
#> Column type frequency:           
#>   character                1     
#>   factor                   2     
#>   numeric                  16    
#> ________________________         
#> Group variables            None  
#> 
#> ── Variable type: character ────────────────────────────────────────────────────
#>   skim_variable   n missing min max empty n_unique whitespace
#> 1 cg_activities 101       0   1 355     0       99          0
#> 
#> ── Variable type: factor ───────────────────────────────────────────────────────
#>   skim_variable   n missing ordered n_unique top_counts                       
#> 1 adopter       101       0 FALSE          4 I t: 63, I w: 22, I t: 15, I o: 1
#> 2 sd_gender     101       0 FALSE          3 mal: 60, fem: 40, div: 1, [no: 0 
#> 
#> ── Variable type: numeric ──────────────────────────────────────────────────────
#>    skim_variable              n missing min  q1 median   q3  max   mean      sd
#>  1 case                     101       0 170 542    936 1199 1549 880.   385.   
#>  2 use_private              101       0   1   2      2    3    5   2.59   1.01 
#>  3 use_work                 101       0   1   1      2    4    5   2.41   1.39 
#>  4 cg_adoption_advantage_01 101       0  -9   3      4    4    5   3.21   2.02 
#>  5 cg_adoption_advantage_02 101       0   1   2      3    4    5   2.69   1.21 
#>  6 cg_adoption_advantage_03 101       0   1   3      4    4    5   3.58   1.07 
#>  7 cg_adoption_advantage_04 101       0   1   3      4    4    5   3.53   0.996
#>  8 cg_adoption_fearofuse_01 101       0   1   2      3    4    5   3.12   1.13 
#>  9 cg_adoption_fearofuse_02 101       0   1   2      3    4    5   3.07   1.15 
#> 10 cg_adoption_fearofuse_03 101       0  -9   3      3    4    5   3.08   1.59 
#> 11 cg_adoption_fearofuse_04 101       0   1   2      3    3    5   2.73   1.09 
#> 12 cg_adoption_social_01    101       0  -9   2      2    3    5   2.39   1.60 
#> 13 cg_adoption_social_02    101       0   1   1      2    3    5   2.36   1.16 
#> 14 cg_adoption_social_03    101       0  -9   1      2    3    5   1.88   1.59 
#> 15 cg_adoption_social_04    101       0   1   1      2    3    5   2.33   1.23 
#> 16 sd_age                   101       0  18  27     38   52   68  39.7   13.8  
#>    ci.low ci.high items alpha
#>  1 804.    956.   NA    NA   
#>  2   2.39    2.79 NA    NA   
#>  3   2.13    2.68 NA    NA   
#>  4   2.81    3.61 NA    NA   
#>  5   2.45    2.93 NA    NA   
#>  6   3.37    3.80 NA    NA   
#>  7   3.34    3.73 NA    NA   
#>  8   2.90    3.34 NA    NA   
#>  9   2.84    3.30 NA    NA   
#> 10   2.77    3.39 NA    NA   
#> 11   2.52    2.95 NA    NA   
#> 12   2.07    2.70 NA    NA   
#> 13   2.13    2.59 NA    NA   
#> 14   1.57    2.19 NA    NA   
#> 15   2.08    2.57 NA    NA   
#> 16  37.0    42.4  NA    NA